Add Huge Update

+++ ShopDemoP/src/main/java/com/demop/ShopDemoP/domain/User/Customer.java
@@ -1,7 +1,7 @@
 package com.demop.ShopDemoP.domain.User;
 
 public class Customer extends User{
-    public Customer(Integer id, String password, String email, String name) {
+    public Customer(Long id, String password, String email, String name) {
         super(id, password, email, name);
     }
 }
