+++ ShopDemoP/src/main/java/com/demop/ShopDemoP/domain/User/Seller.java
@@ -1,7 +1,7 @@
 package com.demop.ShopDemoP.domain.User;
 
 public class Seller extends User {
-    public Seller(Integer id, String password, String email, String name){
+    public Seller(Long id, String password, String email, String name){
         super(id, password, email, name);
     }
 }
